requirements.txt
Flask==2.2.3
Pillow==9.5.0
reportlab==3.6.12
numpy==1.24.3
gunicorn